# few minor MemoryCache perf improvements