I just took a quick look at the source code to see if we could get some gains in the Caching TechEmpower benchmark. This PR is a free lunch, there is definitely more low hanging fruits there and I hope that it becomes one of our .NET 6 goals.

The performance repo microbenchmarks (https://github.com/dotnet/performance/pull/1598) show 5 to 15% perf improvement

|          Method |           Toolchain |      Mean | Ratio |
|---------------- |-------------------- |----------:|------:|
|          GetHit |  \after\CoreRun.exe | 137.61 ns |  0.86 |
|          GetHit | \before\CoreRun.exe | 159.17 ns |  1.00 |
|                 |                     |           |       |
|  TryGetValueHit |  \after\CoreRun.exe | 131.57 ns |  0.84 |
|  TryGetValueHit | \before\CoreRun.exe | 157.07 ns |  1.00 |
|                 |                     |           |       |
|         GetMiss |  \after\CoreRun.exe |  98.56 ns |  0.87 |
|         GetMiss | \before\CoreRun.exe | 113.47 ns |  1.00 |
|                 |                     |           |       |
| TryGetValueMiss |  \after\CoreRun.exe |  97.62 ns |  0.86 |
| TryGetValueMiss | \before\CoreRun.exe | 113.05 ns |  1.00 |
|                 |                     |           |       |
|     SetOverride |  \after\CoreRun.exe | 362.80 ns |  0.95 |
|     SetOverride | \before\CoreRun.exe | 381.01 ns |  1.00 |

The TechEmpower benchmark shows +6k RPS (from 234,210 240,778) to which translates to +2.75% RPS