@kunalspathak 2020-11-24
Perf lab data showing improvements:

https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/master_x64_Windows%2010.0.18362/Microsoft.Extensions.Caching.Memory.Tests.MemoryCacheTests.GetHit.html

https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/master_x64_Windows%2010.0.18362/Microsoft.Extensions.Caching.Memory.Tests.MemoryCacheTests.GetMiss.html