@kunalspathak 2020-11-24
Related improvement link: https://github.com/DrewScoggins/performance-2/issues/3516